Check overloads against existing overloads and macros for ambiguity. (#16)
* Check overloads against existing overloads and macros for ambiguity.
This implementation pretty much follows the Java one.Note that I did not
add a source location to the error. This seems to be inline with the
Java implementation. If we wanted to add this, we would probably have to
pass the environment the sourceinfo which would lead to more changes

* Remove unused line + make inserting a declaration clearer

* Take care of review comments. In particular add formatFunction.

* fix import in types.go

* No need for explicit param in the anonymous func