CelService server for cel-go.
Use Go proto libraries exported from cel-spec.

Note that we use a different Go package for the google.rpc.Status message
than is specified in its .proto file.  We do this to avoid a different
and incompatible Go proto library that is brought in by the Bazel rules.
We will revert to the right package when this bug is fixed.
See https://github.com/bazelbuild/rules_go/issues/1530

Expand common.Source to subsume interpreter.Metadata, for the latter's
eventual deprecation, and to allow the checker inputs to be constructed
without access to the source text.

Introduce a specialized traits.Lister for arrays of ref.Value values.

Implement CelService server and tests.  Includes conversion between
ref.Value types and protos.