The following pull request introduces a unified type provider which takes a list of supported message types and makes them available to the checker and the interpreter. The unified model makes it easier to ensure that the same types are used in both stages of execution, which should also improve object re-use and clarify the execution contract.
This change also introduces an extensive set of type implementations. The
types radically depart from the former model of interpreter types and makes
it simple for developers to introduce their own types that supply operator
overloads and participate in function dispatch.

In terms of pending changes, the implementation does not yet handle well-
known protobuf types such as google.protobuf.Value, google.protobuf.Any
and the associated wrapper types. Support for Value and Any should be
relatively straight forward to implement, though the wrapper integration
may take a bit more work as the type of a wrapper is distinct from the
implicit type conversions for Any and Value.