Fix nil pointer issue in LogicalOr (#111)
Fixed type test logic within short-circuiting paths

The AST Walker generates jump instructions when values are equal to the jump condition.
It is possible that the jump to the call site may occur when not all arguments have values
as it has been predetermined by the walker that the values are not necessary to the eval
result. However, the `nil`-able arguments were being referenced like full values and not
appropriately checked before use in a couple of key locations; thus, resulting in panics.

A `nil` value may occur during message construction when the type referenced cannot be
found, but this is a separate issue and will be dealt with in another pull request.