Fix for SelectExpr field resolution on Unknown/Error types. (#113)
When an unknown value appears as the operand to a SelectExpr, e.g. a.b where a is
unknown, the eval logic attempts to resolve the whole expression a.b to a namespaced
field if possible. In circumstances where this was not possible, the eval behavior was leaving
the select state as nil. This scenario can cause problems further in evaluation, so the code
has been modified to more appropriately handle this case.