Use builtin "context" now that it is standard Go.
This was taken from a Google-internal large scale change, making it to
the upstream repo instead. The originating change said:

The "context" library added in Go 1.7 was based on Google's and has
the same API. It is now aliased to "context" so the two can be used
interchangeably. This change is a cleanup that standardizes on the
standard library "context" package.