Fix bazel breakages by upgrading bazel rules.
The cel-go repository wasn't building for recent (0.20.0)
versions of Bazel, due to assumptions in the older versions
of our imported bazel rules for golang and gRPC.

The Bazel support for gRPC is transitioning from
    https://github.com/pubref/rules_protobuf
to
    https://github.com/stackb/rules_proto
and both overlap with the functionality provided by
    https://github.com/bazelbuild/rules_go/
and none of the above play nicely with
    https://github.com/googleapis/googleapis

Some changes are in the works from the above.  In the meantime,
it looks like we can minimize our dependencies by importing the
gRPC libraries explicitly.