Recursive instruction planner (#157)
The recursive instruction planner walks the recursive AST structure to
create a call graph that executes linearly.

This implementation replaces the AST walker which generated Instruction
objects assembled into a Program. The former implementation had a high
overhead for small expressions as it required a per-invocation memory
allocation to hold program state in a thread-safe manner. The current
implementation does not suffer from this drawback and executes twice
as fast for common code paths.

The following improvements have also been made to address issue #159:
* Isolate state between concurrent evaluations of an Interpretable.
* Switch from pessimistic error checks to optimistically valued checks.
* Optimize constant references and expressions within CEL Programs
* First class functional options for Program flags and improved constant expression computation.
* Recursive instruction planner.

A bug fix has also been applied for issue #155 which has revealed a bit of ugliness with respect
to enforcing homogenous type / value equality within dynamic lists and maps.
* Update code for conformance equality fix.
* Additional type comparison changes for aggregate values.

Closes #159, #155