Add additional tests from C++ parser
When working on a C++ parser, we had several other tests for some
features, such as proper number type assignments in expressions, unicode
encoding, negation (boolean and numeric) and fixed numbered arguments.
We thought it would be useful to port them to the Go parser in order to
have the same test suites on both sides (and they might help the next
person to write a parser).