Updates the has() macro to be more consistent with the cel-spec. (#332)
The cel-spec specifies the following behavior for the has() macro:

```
This means means that has(e.f) ... produces false if f is declared but not set
(or, in proto3, has its default value). Moreover, has(e.f) where e is a map
returns false if f is not defined in the map.
```

In proto2 all non-repeated fields are represented by pointers which will be nil if not set. In proto3 all message fields are represented by pointers which will be nil if not set. The has() macro must also return false if the proto3 primitive field is equal to its zero value.

In the process of making cel-go more consistent with cel-spec, it seems worth reexamining whether the proto3 test for default-ness should also be extended to repeated types (array, map) for both proto2 and proto3 where:

```
has(e.repeated_f) returns false if repeated_f contains zero elements. 
```

The code has been implemented to reflect this slight deviation from the cel-spec, and an issue for google/cel-spec#113 to address the issue at the spec level.