Add AST node IDs to types.Err for errorable expressions
Currently, it is not possible to obtain the source location of a runtime
error making it difficult sometimes to identify the cause of an issue.
If an AST node ID is available, the source location can be obtained, so
add this information to runtime errors.

ref.Val values returned by the following types are examined for error
state and if an error and no AST node has been set, the node ID of the
Interpretable is added to the error.

* evalAnd
* evalAttr
* evalBinary
* evalExhaustiveConditional
* evalList
* evalMap
* evalMap
* evalObj
* evalOr
* evalTestOnly
* evalUnary
* evalVarArgs
* evalWatchAttrQual
* evalWatchConstQual
* evalWatchQual
* evalZeroArity