Fix incorrect example in string formatting docs.
%s is defined to support "all numerical types (int, uint, and double)."
but later the docs also say:
"Passing an incorrect type (an integer to `%s`) is considered an error"

The example in the parenthetical contradicts the definition of %s.

This change replaces the parenthetical example with one that is
accurate and verified by an existing test case. This change also adds
an explicit test case to verify that it is valid to pass an in to %s.