Fix issue related to native byte arrays in type system
Byte arrays are now always considered bytes by both the type checker and
the runtime. This fixes a bug, when there was a native struct containing
an byte array and a function that takes the byte arrays as input, the
program would fail at runtime, as it would expect a dynamic list instead
of bytes.

For an concrete example see the TestFunctionOverloadForNative test in
the native tests.

Fixes #953