Overhaul of policy compilation rules
Better support for fall-through on optional.none() cases from parent rules
as the notion of a `none` value should signify that no branch condition is
satisfied, and thus other branches should be able to assess the rule.

In some cases the `orValue` usage is a little unusual, in that it can
wrap a ternary containing an optional return value, but such cases only
occur when the rule always returns an output and they behave correctly.