Add option to use inaccessible accumulator var (#1097)
Add an option to use '@result' as the accumulator variable for builtin
comprehensions. The current default "__result__" is accesible in the CEL
Source, allowing for expressions to type check but lead to unexpected or
incorrect results. '@result' isn't a normally accessible identifier in the
source expression so a bit safer as a default.