Add LateFunctionBinding declaration and fix constant folding
Adds the declaration for LateFunctionBindings, which can be used to
indicate to the Runtime and the Optimizers that the function will
be bound at runtime through the Activation.

This lets the constant folding optimizer know that the function
potentially has side effects and cannot be folded. Without this
the optimization will fail with an error for late bound functions
where all arguments are constants.

The implementation for late bound functions will be added in a
subsequent commit.