Add support for global constants to constant folder
Global constants like enums are represented in the AST as identifiers. These
identifiers can be looked up in the AST reference map. If there is an entry
in the reference map, we can try to fold it as a constant and to turn it into
a literal.

This reduces the size of the AST and unlocks more optimization
opportunities.