Add documentation for YAML quirks in celpolicy (#1186)
Also add top-level description field.

The celpolicy parser has special handling for CEL expressions embedded
in the YAML document in order to preserve location information. This is
used to help the CEL compiler correctly highlight error location.

As a consequence, the CEL expression will include
characters that are normally not considered meaningful in YAML, and the
literal and folded style for multiline strings are treated identically.