Adding ability to use JSON Names for pb fields
This adds a simple variable to map the fields to the JSONName() instead of
Name() on the protobuf field descriptors.

This allows for using cel expressions against the same JSON
representations in some scenarios.

The default behavior is identical. Test has been added to validate field
names with setting turned on.